v0.1
currently working on the table waitlist tab.

login.vb
-condensed code for mysql queries

tables.designer.vb
-added text box for name in waitlist tab
-added text box for table size in waitlist tab
-added button to add a name and table size to the waitlist
-extra textbox for debugging whatever stuff you want right now

tables.vb
-fixed code for waiter4label
-button for adding a name and table size is finished. code is "secure"
for adding people to the waitlist
-deleting people from the waitlist "works" but not correctly

things that need finished
-need to correctly delete people from the waitlist. what needs to happen
is whatever row that is selected in the DataGridView object needs
deleted and no other row, even if other rows have duplicate information.
currently the commented-out code retrieves the correct row information
but how to use it is still being worked on. or i could try to figure out
how to choose a specific row in the table by a name somehow, i'm not
sure.