v0.12
completed waitlist

login.designer.vb
-changed appearance of number buttons

login.vb
-clear button now exits the program (not supposed to do this but it is
easier to debug the application)

tables.designer.vb
-added a keyboard, includes only letters and numbers (numbers are
located above letter keys). they are in the same positions as they would
be on a physical keyboard
-added labels for the datagridview
-added clear button for the textboxes
-added an error label
-adjusted datagridview size
-moved textboxes around

tables.vb
-keyboard buttons work properly. whenever a textbox is selected, a flag
is set so that the keyboard buttons know where to send a character to.
on default the flag is set to name so that the characters will be sent
to the name textbox
-datagridview changes to not view column1 of the table as well as
adjusts the column sizes to fit more space (this mightve been added in
the v0.11?)
-error label is not visible on default, but if either the name or table
size is blank it will give an error. if table size is not a number it
will give a different error. both of these errors will not open any
connections to the database